relocate included files due to pathing errors.
Unfortunately cannot keep v8 header files inside a v8 folder.
Added a static cast to v8-platform.h to suppress warning.

+++ include/v8-platform.h
@@ -1130,7 +1130,7 @@ class Platform {
    * required.
    */
   virtual int64_t CurrentClockTimeMilliseconds() {
-    return floor(CurrentClockTimeMillis());
+    return static_cast<int64_t>(floor(CurrentClockTimeMillis()));
   }
 
   /**
